+++ src/components/loader.js
@@ -14,7 +14,7 @@ const Loader = ({ onLoaderLoaded }) => {
 
     const { t: tr } = useTranslation();
     const appContext = useContext(AppContext);
-    const { apiPath, setApiPath, setApiVersion, vtcLogo, setVtcLogo, vtcBanner, setVtcBanner, vtcBackground, setVtcBackground, setSpecialRoles, setSpecialUsers, setPatrons, setApiConfig, webConfig, setWebConfig, loadLanguages, setAllRoles, setAllPerms, setAllRanks, loadMemberUIDs, loadDlogDetails } = useContext(AppContext);
+    const { apiPath, setApiPath, setApiVersion, vtcLogo, setVtcLogo, vtcBanner, setVtcBanner, vtcBackground, setVtcBackground, setSpecialRoles, setSpecialUsers, setPatrons, setUserConfig, setApiConfig, webConfig, setWebConfig, loadLanguages, setAllRoles, setAllPerms, setAllRanks, loadMemberUIDs, loadDlogDetails } = useContext(AppContext);
     const { themeSettings, setThemeSettings } = useContext(ThemeContext);
 
     const [isMember, setIsMember] = useState(false);
@@ -172,7 +172,7 @@ const Loader = ({ onLoaderLoaded }) => {
                 setPatrons(patrons);
             }
             if (userConfig) {
-                vars.userConfig = userConfig;
+                setUserConfig(userConfig);
             }
             if (config) {
                 setApiConfig(config.config);
@@ -201,7 +201,7 @@ const Loader = ({ onLoaderLoaded }) => {
                 writeLS("cache", cache, window.dhhost + webConfig.abbr + webConfig.api_host);
             }
 
-            let auth = await FetchProfile({ ...appContext, apiPath: apiPath, webConfig: webConfig, specialusers: specialUsers, patrons: patrons });
+            let auth = await FetchProfile({ ...appContext, apiPath: apiPath, webConfig: webConfig, specialUsers: specialUsers, patrons: patrons });
             setIsMember(auth.member);
 
             setThemeSettings(prevSettings => ({ ...prevSettings })); // refresh theme settings