+++ src/variables.js
@@ -1,5 +1,3 @@
-var userConfig = {};
-exports.userConfig = userConfig;
 var defaultUserLevel = 0; // this affects all users;
 exports.defaultUserLevel = defaultUserLevel;
 var vtcLevel = 0; // 0: regular / 1: premium / 3: special